fix CI/CD issues
