+++ .github/workflows/ci.yml
@@ -77,7 +77,7 @@ jobs:
         run: pnpm install --frozen-lockfile
 
       - name: Lint
-        run: pnpm run lint
+        run: pnpm run lint -- --max-warnings=100
 
   format-check:
     name: Format Check
