+++ src/app/opengraph-image.tsx
@@ -1,5 +1,7 @@
 import { ImageResponse } from 'next/og';
 
+export const runtime = 'edge';
+
 // Image metadata
 export const alt = 'StrellerMinds - Blockchain Education Platform';
 export const size = {
